Handle response format change
The Bento API is responding without the data wrapper key when making command changes. The response validation was explicitly checking for the 'data' key and throwing an exception if it was not present. I removed this check and added a string cast due to an error experienced in the Lumen framework.